pyproject.toml
[tool.poetry]
name = "doccron"
version = "1.6.0-alpha.0"
description = "Schedule with Docstrings"
authors = ["Ronie Martinez <ronmarti18@gmail.com>"]
license = "MIT"
repository = "https://github.com/roniemartinez/DocCron"
readme = "README.md"
keywords = [
    "crontab",
    "cron",
    "docstring",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
]

[tool.poetry.dependencies]
python = ">=3.6.2, <4"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
autoflake = "^1.3.1"
black = "^21.9b0"
codecov = "^2.0.16"
dephell = "^0.8.3"
flake8 = "^4.0.1"
freezegun = "^1.1.0"
isort = "^5.9.3"
mypy = "^0.910"
pyproject-flake8 = "^0.0.1-alpha.2"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
tomlkit = "0.7.0"  # dephell is broken on versions above 0.7.0
types-freezegun = "^1.1.3"
types-python-dateutil = "^2.8.2"

[tool.isort]
line_length = 120
multi_line_output = 3
force_grid_wrap = 0
use_parentheses = true
include_trailing_comma = true
ensure_newline_before_comments = true
atomic = true

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = """
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/setup.py
"""

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "pkg_resources",
]
ignore_missing_imports = true

[tool.flake8]
max-line-length = 120
extend-ignore = "E203"
extend-exclude = """
setup.py,
"""

[tool.pytest.ini_options]
addopts = "--cov=doccron --cov-report term-missing --cov-report=xml --cov-report=html -vv"

[tool.dephell.main]
from = { format = "poetry", path = "pyproject.toml" }
to = { format = "setuppy", path = "setup.py" }

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
